#7 - Refactored e2e
