+++ e2e/index.spec.ts
@@ -1,33 +1,18 @@
 import { Browser, devices } from 'puppeteer';
-import { configureToMatchImageSnapshot } from 'jest-image-snapshot';
-import 'expect-puppeteer';
-import fs from 'fs';
+import { mkdir, timeout, url, customSnapshotsDir, customDiffDir } from './setup';
 
 declare const browser: Browser;
 
-const customConfig = { threshold: 0.5 };
-const toMatchImageSnapshot = configureToMatchImageSnapshot({
-  customDiffConfig: customConfig,
-});
-
-expect.extend({ toMatchImageSnapshot });
-
-const timeout = 10000;
-const url = 'http://127.0.0.1:7683/ts-react-playground/';
-
-const mkdir = (path: string) => {
-  if (!fs.existsSync(path)) fs.mkdirSync(path);
-};
-
 describe('Index', () => {
   beforeAll(async () => {
-    mkdir('./it/screenshots');
+    mkdir(customSnapshotsDir);
+    mkdir(customDiffDir);
   });
 
   Object.keys(devices).forEach((key) => {
     const device = devices[key];
 
-    xtest(
+    test(
       `${device.name}`,
       async () => {
         const page = await browser.newPage();
@@ -44,8 +29,7 @@ describe('Index', () => {
         await page.goto(url);
         await page.waitForSelector('.w3-margin');
 
-        mkdir('./it/screenshots/index/');
-        const image = await page.screenshot({ path: `./it/screenshots/index/${device.name}.png`, type: 'png' });
+        const image = await page.screenshot();
         expect(image).toMatchImageSnapshot();
 
         page.close();
