+++ jest-puppeteer.config.js
@@ -1,3 +1,5 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const webpackConfig = require('./e2e/webpack.config.js');
 module.exports = {
   browser: 'chromium',
   browserContext: 'default',
@@ -6,8 +8,8 @@ module.exports = {
     headless: process.env.HEADLESS !== 'false',
   },
   server: {
-    command: `npm run dev`,
-    port: 7683,
+    command: `npm run e2e:serve`,
+    port: webpackConfig.devServer.port,
     launchTimeout: 10000,
     debug: false,
   },
