+++ package.json
@@ -2,21 +2,16 @@
   "name": "ts-react-playground",
   "version": "0.0.2",
   "description": "TypeScript React Playground",
-  "main": "dist/index.js",
   "scripts": {
-    "dev": "webpack-dev-server --config webpack.config.js",
-    "clear": "rimraf junit.xml ./it/screenshots/index ./node_modules ./dist ./coverage ./coverage-it ./package-lock.json ./public/vendor ./public/script ./public/styles.css .jest-test-results.json",
-    "clear:snapshots": "rimraf ./**/__snapshots__",
+    "start": "webpack-dev-server --open --progress --config webpack.config.js",
+    "clear": "rimraf ./node_modules ./dist ./coverage ./coverage-e2e ./package-lock.json ./public/vendor ./public/script ./public/styles.css .jest-test-results.json junit.xml",
     "livereload": "livereloadx -a ./public",
     "build": "npm run vendor && npm run script && npm run scss",
     "postbuild": "ts-node minify.ts",
     "test": "jest --runInBand --env=enzyme --setupFilesAfterEnv=jest-enzyme --reporters=jest-spec-reporter jest-junit --coverage --coverageDirectory ../coverage --rootDir ./src",
     "test:update": "npm run test -- -u",
-    "preit": "rimraf public-it && node prepareit.js",
-    "it": "jest --runInBand --preset=jest-puppeteer --coverage --coverageDirectory ../coverage-it --rootDir ./it",
-    "postit": "rimraf public-it",
-    "preesmoduleshims": "rimraf ./public/es-module-shims.js",
-    "esmoduleshims": "tsc --project ./src/vendor/es-module-shims",
+    "e2e:serve": "webpack-dev-server --silent --quiet --liveReload false --inline false --config ./e2e/webpack.config.js",
+    "e2e": "jest --runInBand --preset=jest-puppeteer --coverage --coverageDirectory ../coverage-e2e --rootDir ./e2e",
     "prescript": "rimraf ./public/script",
     "script": "tsc --project ./src/script --outDir ./public/script",
     "prewatch-script": "npm run script",
@@ -25,14 +20,14 @@
     "scss": "node-sass --output-style compressed --output ./public ./src/style/styles.scss",
     "prewatch-scss": "npm run scss",
     "watch-scss": "node-sass --output-style compressed  --watch --output ./public ./src/style/styles.scss",
-    "prevendor": "rimraf ./public/vendor",
-    "vendor": "rimraf ./public/vendor && npm run vendor:base && npm run vendor:faker && npm run vendor:fontawesome && npm run vendor:react && npm run vendor:recharts",
+    "prevendor": "rimraf ./public/vendor ./public/es-module-shims.js",
+    "vendor": "rimraf ./public/vendor && npm run vendor:base && npm run vendor:faker && npm run vendor:fontawesome && npm run vendor:react && npm run vendor:recharts && npm run vendor:esmoduleshims",
     "vendor:base": "cross-env NODE_ENV=production rollup -c ./src/vendor/rollup-base.config.js",
     "vendor:faker": "cross-env NODE_ENV=production rollup -c ./src/vendor/rollup-faker.config.js",
     "vendor:fontawesome": "cross-env NODE_ENV=production rollup -c ./src/vendor/rollup-fontawesome.config.js",
     "vendor:react": "cross-env NODE_ENV=production rollup -c ./src/vendor/rollup-react.config.js",
     "vendor:recharts": "cross-env NODE_ENV=production rollup -c ./src/vendor/rollup-recharts.config.js",
-    "postvendor": "npm run esmoduleshims",
+    "vendor:esmoduleshims": "tsc --project ./src/vendor/es-module-shims",
     "docker:build": "cross-env NODE_ENV=production docker-compose build --force-rm --no-cache --parallel --pull",
     "docker:up": "docker-compose up -d",
     "docker:down": "docker-compose down",
@@ -68,6 +63,8 @@
   "jest": {
     "collectCoverage": true,
     "collectCoverageFrom": [
+      "!setup.ts",
+      "!webpack.config.js",
       "!**/tsconfig.json",
       "!**/tsconfig.*.json",
       "!**/*.tsx.snap",
